file sh

+++ setup.sh
@@ -1,73 +1,52 @@
 #!/bin/sh
 
-# Delete YAML
 if [ "$1" = "delete" ]
 then
+	if ! minikube status >/dev/null 2>&1
+		then
+   		echo "\033[1;33mMinikube is not started!\033[0m\n"
+		exit 1
+	fi
     echo "\n\033[0;31mDeleting pods !\033[0m\n"
     kubectl delete -k srcs
 	exit 1
 fi
 
-# Print dashboard
 if [ "$2" = "only" ]
 then
-	MINIKUBE_IP=$(minikube ip)
-	# Print ssh serv
+	if ! minikube status >/dev/null 2>&1
+	then
+   		echo "\033[1;33mMinikube is not started!\033[0m\n"
+		exit 1
+	fi
     if [ "$1" = "ssh" ]
     then
-        echo "\n\033[0;35mSSH :\n\033[0m"
-        ssh -v root@$MINIKUBE_IP -p 21
+        sh srcs/sh/ssh.sh
     fi
-	# Print minikube
     if [ "$1" = "minikube" ]
     then
-        echo "\n\033[0;35mMINIKUBE DASHBOARD :\n\033[0m"
-        minikube dashboard
+        sh srcs/sh/minikube.sh
     fi
-    # Print dashboard
     if [ "$1" = "dashboard" ]
     then
-        clear
-        echo "\n\n\033[0;35mDASHBOARD :\n\033[0m"
-        kubectl get all
+        sh srcs/sh/dashboard.sh
     fi
-    # Print ip
     if [ "$1" = "ip" ]
     then
-        echo "\n\n                  Your ip is : \033[1;32m http://$MINIKUBE_IP \033[0m\n"
+    	sh srcs/sh/ip.sh
     fi
-    # Print Password
     if [ "$1" = "password" ]
     then
-		echo "\033[1;35m\n  |================|\033[0m"
-		echo "\033[1;35m  |--  PASSWORD  --|\033[0m"
-		echo "\033[1;35m  |________________|\033[0m"
-        echo "\033[0;35m\n  --  \033[0;33mWordpress \033[0;35m--\n\033[0m"
-        echo "\033[0;31madmin  \033[2;35m>\033[0m"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32municorn\033[0m'  \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32municorn\033[0m'\n"
-        echo "\033[0;34muser   \033[2;35m>\033[0m"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mjacens\033[0m'   \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32m123\033[0m'"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mnorminet\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword123\033[0m'"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mesteban\033[0m'  \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mmotdepasse\033[0m'\n"
-        echo "\033[0;35m\n  --    \033[0;33mMySQL    \033[0;35m--\n\033[0m"
-        echo "\033[0;31madmin  \033[2;35m>\033[0m"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mroot\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword\033[0m'\n"
-        echo "\033[0;34muser   \033[2;35m>\033[0m"
-        echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32muser\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword\033[0m'\n"
-        echo "\033[1;35m  -----------------\n\033[0m"
+		sh srcs/sh/password.sh
     fi
     exit 1
 fi
-
-# Ensure docker and minikube are installed
 if ! which docker >/dev/null 2>&1 ||
     ! which minikube >/dev/null 2>&1
 then
     echo "\033[1;31mPlease install Docker and Minikube\033[0m"
     exit 1
 fi
-
-# Ensure minikube is launched
 if ! minikube status >/dev/null 2>&1
 then
     echo "\033[1;33mMinikube is not started! Starting now...\033[0m\n"
@@ -81,75 +60,28 @@ then
     minikube addons enable ingress
 fi
 
-# Building images
-MINIKUBE_IP=$(minikube ip)
-
-cp srcs/ftps/entrypoint srcs/ftps/entrypoint-target
-sed -i '' "s/##MINIKUBE_IP##/$MINIKUBE_IP/g" srcs/ftps/entrypoint-target
-cp srcs/wordpress/wordpress_dump.sql srcs/wordpress/wordpress_dump-target.sql
-sed -i '' "s/##MINIKUBE_IP##/$MINIKUBE_IP/g" srcs/wordpress/wordpress_dump-target.sql
-
+IP=$(minikube ip)
+sh srcs/sh/config_file.sh
 echo "\n\033[1;34mStarting Services\033[0m\n"
 eval $(minikube docker-env)
-
-echo "\n\033[0;36mStarting nginx...\033[0m\n"
-docker build -t custom-nginx:1.11 srcs/nginx
-echo "\n\033[0;36mStarting ftps...\033[0m\n"
-docker build -t custom-ftps:1.6 srcs/ftps
-echo "\n\033[0;36mStarting wordpress...\033[0m\n"
-docker build -t custom-wordpress:1.9 srcs/wordpress
-echo "\n\033[0;36mStarting phpmyadmin...\033[0m\n"
-docker build -t custom-phpmyadmin:1.1 srcs/phpmyadmin
-echo "\n\033[0;36mStarting grafana...\033[0m\n"
-docker build -t custom-grafana:1.0 srcs/grafana
-echo "\n\033[0;36mStarting mysql...\033[0m\n"
-docker build -t custom-mysql:1.11 srcs/mysql
-echo "\n\033[0;34mStarting Services\033[0m"
-echo "\n\033[1;32mOk !\033[0m\n"
-
-# Appli YAML
+sh srcs/sh/build_image.sh
 echo "\n\033[0;35mStarting pods !\033[0m\n"
 kubectl apply -k srcs
 
-# Print dashboard
 if [ "$1" = "dashboard" ]
 then
-    clear
-    echo "\n\n\033[0;35mDASHBOARD :\n\033[0m"
-    kubectl get all
+    sh srcs/sh/dashboard.sh
 fi
-
-# Print Password
 if [ "$1" = "password" ]
 then
-	echo "\033[1;35m\n  |================|\033[0m"
-	echo "\033[1;35m  |--  PASSWORD  --|\033[0m"
-	echo "\033[1;35m  |________________|\033[0m"
-    echo "\033[0;35m\n  --  \033[0;33mWordpress \033[0;35m--\n\033[0m"
-	echo "\033[0;31madmin  \033[2;35m>\033[0m"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32municorn\033[0m'  \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32municorn\033[0m'\n"
-	echo "\033[0;34muser   \033[2;35m>\033[0m"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mjacens\033[0m'   \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32m123\033[0m'"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mnorminet\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword123\033[0m'"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mesteban\033[0m'  \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mmotdepasse\033[0m'\n"
-	echo "\033[0;35m\n  --    \033[0;33mMySQL    \033[0;35m--\n\033[0m"
-	echo "\033[0;31madmin  \033[2;35m>\033[0m"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32mroot\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword\033[0m'\n"
-	echo "\033[0;34muser   \033[2;35m>\033[0m"
-	echo "\033[0;90m  - \033[0;36mname\033[0m \033[2;49m:\033[0m '\033[0;32muser\033[0m' \033[0;36mpwd\033[0m \033[2;49m:\033[0m '\033[0;32mpassword\033[0m'\n"
-	echo "\033[2;35m  -----------------\n\033[0m"
+	sh srcs/sh/password.sh
 fi
-
-echo "\n\n                  Your ip is : \033[1;32m http://$MINIKUBE_IP \033[0m\n"
-# Print minikube
+sh srcs/sh/ip.sh
 if [ "$1" = "minikube" ]
 then
-	echo "\n\n\033[0;35mMINIKUBE DASHBOARD :\n\033[0m"
-	minikube dashboard
+	sh srcs/sh/minikube.sh
 fi
-# Print ssh serv
-    if [ "$1" = "ssh" ]
-    then
-        echo "\n\033[0;35mSSH :\n\033[0m"
-        ssh -v root@$MINIKUBE_IP -p 21
-    fi
+if [ "$1" = "ssh" ]
+then
+    sh srcs/sh/ssh.sh
+ fi
